[Chore #163801300] Separate models from views.
Separate concerns by separating models from views
to make it easier to navigate the folder structure.
