+++ app/tests/v1/test_views/test_party_views.py
